Make the process info cache size configurable
For performance reasons, the process info cache uses a size threshold.
If the system pid_max is <= the process info cache size, an array will
without locking will be used; otherwise, a map with locking will be
used. The threshold was hard-coded to 32768, which is the default stock
kernel value for pid_max. This also increases that default to 131072,
which appears to be the default with Ubuntu server installs.