Adjust the method by which offsets are computed
Each CPU has its own ring buffer from which perf events are pulled.
Each CPU has its own clock that is used to timestamp perf events.
A "best effort" is made by EventMonitor to funnel events from each
CPU in a consistent order. This will never be 100%, but the way in
which offsets were computed prior to this change was terrible. The
clocks for each CPU are now offset relative to CPU 0 using the
kernel's high resolution timer.

Beginning with 4.1 kernels, the clock to use can be selected, with
one of the options being a system-wide monotonic clock. This would
be the ideal clock to use, but it is not supported by a significant
number of deployed systems. We must be able to function without it.
Note that using the sytstem-wide monotonic clock still would not
guarantee perfect ordering, but it would be the closest possible.