Fix TestSubdirs so that it stops failing so often
The test reads from a channel and then attempts to type assert what it
read before it checks to see if it got anything from the channel, which
ends up performing a type assertion on nil that fails, of course. Fix
the test to check the ok flag before performing the type assertion.