Send an initial GetEventsResponse after subscription
Any non-fatal errors, warnings, or messages are reported back to
the client in the status portion of GetEventsResponse. If there
are none to report, an OK status is sent.

The first GetEventsResponse is always a status only response.
Update the functional tests to use this initial message instead
of waiting an arbitrary amount of time before launching its test
container.