Stabilize ProcessId by using boot id, pid, and start time
If the process is so short-lived that a start time cannot be acquired
from the proc fs, use the timestamp on the sample that caused the
sensor to learn of the process. Since it's short-lived, it won't be
found on a successive scan of the proc fs on a sensor restart anyway,
so the actual timestamp used doesn't really matter as long as the
value chosen cannot be reused again later.