Handle out of order fork events on kernels older than 3.9
Kernels older than 3.9 do not have the task/task_newtask tracepoint,
which gives the sensor everything it needs in one event to handle
task cloning. For kernels lacking this tracepoint, a kprobe and a
different tracepoint must be used, but one depends on the data from
the other. Since events can sometimes come out of order when they
occur on different CPUs, the sensor must be able to handle these
two events coming out of order. The intent was there prior to this
revision, but the work to handle it properly was never completed.
This revision completes it.