Filter kprobe registrations through /proc/kallsyms
The kernel exports a file called kallsyms in the proc fs that lists all
all of the symbols exported by the kernel. Make all kprobe registrations
from the sensor check for symbol existence before trying to register the
kprobe with the kernel. Doing this will remove kernel log noise about
missed kprobes when we're probing for different kernel versions, but
more importantly this is laying the groundwork for future kprobes
that'll be used in the sensor that are more complicated due to kernel
code churn across kernel versions.